Add shutdown support - merge from private branch shutdown_support_trunk. The shutdown support consist of a set of different changes all of them together solving the problem.
1.       Implement atexit() registration propagation and catching across all our DLLs on Windows and Linux
2.       Implement TerminateProcess() procedure
3.       Implement OpenCL API disabling after process termination was discovered while still supporting ICD functionality and current implementation 
4.       Change in SharedPtr mechanism that allows implementation of specific lifecycle per class. 
5.       Add important external classes tracking (events, mem-objects, queues, devices) using the class-specific lifecycle implementation
6.       Add support for queues and tasks cancellation – both in a framework and both devices
7.       Some changes in GenericMemoryObject multi-device subbuffers support to break SharedPtr loops
8.       Some changes in in-order queue implementation – bug fixes to support tasks cancellation
9.       Fixes in OOO queue – fix intermittent crashes and tasks cancellation support
10.   Resurrecting back NotificationPort shutdown support, removed by Adir during windows port of MIC
11.   Many-many-many crash fixes and something I already don’t remember.

+++ opencl-intel/devices/cpu_device/wg_context_pool.cpp
@@ -43,7 +43,7 @@ void WgContextPool::Init(unsigned int maxNumThreads, unsigned int maxNumMasters)
     assert (NULL != m_wgContextPool && "Memory allocation failed");
     assert (NULL != m_wgContextWrapperPool && "Memory allocation failed");
 
-    for (long i = maxNumMasters; i < m_maxNumThreads; ++i)
+    for (unsigned int i = maxNumMasters; i < m_maxNumThreads; ++i)
     {
         m_wgContextWrapperPool[i].pContext = m_wgContextPool + i;
     }
@@ -58,7 +58,7 @@ void WgContextPool::Clear()
 {
     if (NULL != m_wgContextWrapperPool)
     {
-        for (long i = 0; i < m_maxNumThreads; ++i)
+        for (unsigned int i = 0; i < m_maxNumThreads; ++i)
         {
             m_wgContextWrapperPool[i].pContext = NULL;
         }
